+++ packages/medusa/src/helpers/test-request.js
@@ -1,8 +1,4 @@
-import {
-  moduleHelper,
-  moduleLoader,
-  registerModules,
-} from "@medusajs/modules-sdk"
+import { registerModules } from "@medusajs/modules-sdk"
 import { asValue, createContainer } from "awilix"
 import express from "express"
 import jwt from "jsonwebtoken"
@@ -15,6 +11,9 @@ import featureFlagLoader, { featureFlagRouter } from "../loaders/feature-flags"
 import passportLoader from "../loaders/passport"
 import servicesLoader from "../loaders/services"
 import strategiesLoader from "../loaders/strategies"
+import moduleLoader, { moduleHelper } from "../loaders/module"
+import repositories from "../loaders/repositories"
+import models from "../loaders/models"
 
 const adminSessionOpts = {
   cookieName: "session",
@@ -40,7 +39,31 @@ const config = {
 
 const testApp = express()
 
+function asArray(resolvers) {
+  return {
+    resolve: (container) =>
+      resolvers.map((resolver) => container.build(resolver)),
+  }
+}
+
 const container = createContainer()
+
+// TODO: remove once the util is merged in master
+container.registerAdd = function (name, registration) {
+  const storeKey = name + "_STORE"
+
+  if (this.registrations[storeKey] === undefined) {
+    this.register(storeKey, asValue([]))
+  }
+  const store = this.resolve(storeKey)
+
+  if (this.registrations[name] === undefined) {
+    this.register(name, asArray(store))
+  }
+  store.unshift(registration)
+
+  return this
+}.bind(container)
 
 container.register("featureFlagRouter", asValue(featureFlagRouter))
 container.register("modulesHelper", asValue(moduleHelper))
@@ -66,6 +89,8 @@ testApp.use((req, res, next) => {
 })
 
 featureFlagLoader(config)
+models({ container, configModule: config, isTest: true })
+repositories({ container, isTest: true })
 servicesLoader({ container, configModule: config })
 strategiesLoader({ container, configModule: config })
 passportLoader({ app: testApp, container, configModule: config })