+++ packages/medusa/src/repositories/image.ts
@@ -1,8 +1,26 @@
+import { QueryDeepPartialEntity } from "typeorm/query-builder/QueryPartialEntity"
+
 import { Image } from "../models"
 import { dataSource } from "../loaders/database"
 import { In } from "typeorm"
 
 export const ImageRepository = dataSource.getRepository(Image).extend({
+  async insertBulk(data: QueryDeepPartialEntity<Image>[]): Promise<Image[]> {
+    const queryBuilder = this.createQueryBuilder()
+      .insert()
+      .into(Image)
+      .values(data)
+
+    // TODO: remove if statement once this issue is resolved https://github.com/typeorm/typeorm/issues/9850
+    if (!queryBuilder.connection.driver.isReturningSqlSupported("insert")) {
+      const rawImages = await queryBuilder.execute()
+      return rawImages.generatedMaps.map((d) => this.create(d)) as Image[]
+    }
+
+    const rawImages = await queryBuilder.returning("*").execute()
+    return rawImages.generatedMaps.map((d) => this.create(d))
+  },
+
   async upsertImages(imageUrls: string[]) {
     const existingImages = await this.find({
       where: {
@@ -14,16 +32,21 @@ export const ImageRepository = dataSource.getRepository(Image).extend({
     )
 
     const upsertedImgs: Image[] = []
+    const imageToCreate: QueryDeepPartialEntity<Image>[] = []
 
-    for (const url of imageUrls) {
+    imageUrls.forEach((url) => {
       const aImg = existingImagesMap.get(url)
       if (aImg) {
         upsertedImgs.push(aImg)
       } else {
         const newImg = this.create({ url })
-        const savedImg = await this.save(newImg)
-        upsertedImgs.push(savedImg)
+        imageToCreate.push(newImg as QueryDeepPartialEntity<Image>)
       }
+    })
+
+    if (imageToCreate.length) {
+      const newImgs = await this.insertBulk(imageToCreate)
+      upsertedImgs.push(...newImgs)
     }
 
     return upsertedImgs