+++ packages/medusa/src/types/product-variant.ts
@@ -12,8 +12,10 @@ import {
   DateComparisonOperator,
   NumericalComparisonOperator,
   StringComparisonOperator,
+  WithRequiredProperty,
 } from "./common"
 import { XorConstraint } from "./validators/xor"
+import { ProductVariant } from "../models"
 
 export type ProductVariantPrice = {
   id?: string
@@ -82,6 +84,30 @@ export type UpdateProductVariantInput = {
   options?: ProductVariantOption[]
   prices?: ProductVariantPrice[]
   metadata?: Record<string, unknown>
+}
+
+export type UpdateProductVariantData = {
+  variant: ProductVariant
+  updateData: UpdateProductVariantInput
+}
+
+export type UpdateVariantPricesData = {
+  variantId: string
+  prices: ProductVariantPrice[]
+}
+
+export type UpdateVariantRegionPriceData = {
+  variantId: string
+  price: {
+    currency_code: string
+    region_id: string
+    amount: number
+  }
+}
+
+export type UpdateVariantCurrencyPriceData = {
+  variantId: string
+  price: WithRequiredProperty<ProductVariantPrice, "currency_code">
 }
 
 export class FilterableProductVariantProps {