+++ www/docs/src/theme/DocCard/index.js
@@ -9,25 +9,33 @@ import clsx from 'clsx';
 import isInternalUrl from '@docusaurus/isInternalUrl';
 import styles from './styles.module.css';
 import {translate} from '@docusaurus/Translate';
-import ThemedImage from '@theme/ThemedImage';
+import BorderedIcon from '../../components/BorderedIcon';
+import Badge from '../../components/Badge';
+import Icons from '../Icon';
 
 function CardContainer({href, children, className}) {
   return (
-    <Link
-      href={href}
-      className={clsx('card', styles.cardContainer, className)}>
-      {children}
-    </Link>
+    <article className={`card-wrapper margin-bottom--lg`}>
+      <Link
+        href={href}
+        className={clsx('card', styles.cardContainer, className)}>
+        {children}
+      </Link>
+    </article>
   );
 }
-function CardLayout({href, icon, title, description, containerClassName}) {
+
+function CardLayout({href, icon, title, description, html, containerClassName, isSoon = false}) {
   return (
-    <CardContainer href={href} className={containerClassName}>
-      {icon}
+    <CardContainer href={href} className={clsx(containerClassName, isSoon && styles.cardSoon)}>
+      <div className={clsx(styles.cardIconContainer)}>
+        {icon}
+        {isSoon && <Badge variant={'purple'}>Guide coming soon</Badge>}
+      </div>
       <div className={clsx(styles.contentContainer)}>
-        <h2 className={clsx(styles.cardTitle)} title={title}>
+        <span className={clsx(styles.cardTitle)} title={title}>
           {title}
-        </h2>
+        </span>
         {description && (
           <p
             className={clsx(styles.cardDescription)}
@@ -35,12 +43,53 @@ function CardLayout({href, icon, title, description, containerClassName}) {
             {description}
           </p>
         )}
+        {html && (
+          <p
+            className={clsx(styles.cardDescription)}
+            dangerouslySetInnerHTML={{
+              __html: html
+            }}
+          ></p>
+        )}
       </div>
     </CardContainer>
   );
 }
+
+function getCardIcon (item) {
+  if (item.customProps?.themedImage) {
+    return (
+      <BorderedIcon icon={{
+        light: item.customProps.themedImage.light,
+        dark: item.customProps.themedImage.dark
+      }} wrapperClassName='card-icon-wrapper' iconClassName={'card-icon'} />
+    )
+  } else if (item.customProps?.image) {
+    return (
+      <BorderedIcon icon={{
+        light: item.customProps.image
+      }} wrapperClassName='card-icon-wrapper' iconClassName={'card-icon'} />
+    );
+  } else if (item.customProps?.icon) {
+    return <BorderedIcon 
+      IconComponent={item.customProps.icon} 
+      wrapperClassName='card-icon-wrapper' iconClassName={'card-icon'} />;
+  } else if (item.customProps?.iconName && Icons.hasOwnProperty(item.customProps?.iconName)) {
+    return <BorderedIcon
+      IconComponent={Icons[item.customProps?.iconName]}
+      wrapperClassName='card-icon-wrapper' iconClassName={'card-icon'} />;
+  } else {
+    return (
+      <div className={clsx('card-icon-wrapper', 'no-zoom-img')}>
+        {isInternalUrl(item.href) ? '📄️' : '🔗'}
+      </div>
+    );
+  }
+}
+
 function CardCategory({item}) {
   const href = findFirstCategoryLink(item);
+  const icon = getCardIcon(item);
   // Unexpected: categories that don't have a link have been filtered upfront
   if (!href) {
     return null;
@@ -48,11 +97,11 @@ function CardCategory({item}) {
   return (
     <CardLayout
       href={href}
-      icon="🗃️"
+      icon={icon}
       title={item.label}
       description={translate(
         {
-          message: '{count} items',
+          message: item.customProps?.description || '{count} items',
           id: 'theme.docs.DocCard.categoryDescription',
           description:
             'The default description for a category card in the generated index about how many items this category includes',
@@ -60,35 +109,13 @@ function CardCategory({item}) {
         {count: item.items.length},
       )}
       containerClassName={item.customProps?.className}
+      isSoon={item.customProps?.isSoon}
     />
   );
 }
+
 function CardLink({item}) {
-  let icon;
-  if (item.customProps && item.customProps.themedImage) {
-    icon = (
-      <div className={clsx(styles.imageContainer, 'no-zoom-img')}>
-        <ThemedImage alt={item.label} sources={{
-          light: item.customProps.themedImage.light,
-          dark: item.customProps.themedImage.dark
-        }} />
-      </div>
-    )
-  } else if (item.customProps && item.customProps.image) {
-    icon = (
-      <div className={clsx(styles.imageContainer, 'no-zoom-img')}>
-        <img src={item.customProps.image} alt={item.label} />
-      </div>
-    );
-  } else if (item.customProps && item.customProps.icon) {
-    icon = item.customProps.icon;
-  } else {
-    icon = (
-      <div className={clsx(styles.imageContainer, 'no-zoom-img')}>
-        {isInternalUrl(item.href) ? '📄️' : '🔗'}
-      </div>
-    );
-  }
+  const icon = getCardIcon(item);
   const doc = useDocById(item.docId ?? undefined);
 
   return (
@@ -97,10 +124,13 @@ function CardLink({item}) {
       icon={icon}
       title={item.label}
       description={item.customProps?.description || doc?.description}
+      html={item.customProps?.html}
       containerClassName={item.customProps?.className}
+      isSoon={item.customProps?.isSoon}
     />
   );
 }
+
 export default function DocCard({item}) {
   switch (item.type) {
     case 'link':