+++ www/docs/src/theme/Navbar/Content/index.js
@@ -6,12 +6,12 @@ import {
 } from '@docusaurus/theme-common/internal';
 import NavbarItem from '@theme/NavbarItem';
 import NavbarColorModeToggle from '@theme/Navbar/ColorModeToggle';
-import SearchBar from '@theme/SearchBar';
 import NavbarMobileSidebarToggle from '@theme/Navbar/MobileSidebar/Toggle';
 import NavbarLogo from '@theme/Navbar/Logo';
-import NavbarSearch from '@theme/Navbar/Search';
 import styles from './styles.module.css';
 import {useWindowSize} from '@docusaurus/theme-common';
+import NavbarActions from '../Actions';
+import Tooltip from '../../Tooltip';
 
 function useNavbarItems() {
   // TODO temporary casting until ThemeConfig type is improved
@@ -38,7 +38,7 @@ export default function NavbarContent() {
   const mobileSidebar = useNavbarMobileSidebar();
   const items = useNavbarItems();
   const [leftItems, rightItems] = splitNavbarItems(items);
-  const windowSize = useWindowSize();
+  const { navbarActions } = useThemeConfig()
 
   return (
     <NavbarContentLayout
@@ -55,12 +55,11 @@ export default function NavbarContent() {
         // Ask the user to add the respective navbar items => more flexible
         <>
           <NavbarItems items={rightItems} />
-          <NavbarColorModeToggle className={styles.colorModeToggle} />
-          {windowSize === 'mobile' && (
-            <NavbarSearch>
-              <SearchBar />
-            </NavbarSearch>
-          )}
+          <span className='divider'></span>
+          <Tooltip text='Switch theme'>
+            <NavbarColorModeToggle className={styles.colorModeToggle} />
+          </Tooltip>
+          <NavbarActions items={navbarActions} />
         </>
       }
     />