+++ www/docs/src/themes/medusaDocs.js
@@ -3,8 +3,8 @@ const palenightTheme = require("prism-react-renderer/themes/palenight")
 const theme = {
   ...palenightTheme,
   plain: {
-    color: palenightTheme.plain.color,
-    backgroundColor: '#262626'
+    color: '#7E7D86',
+    backgroundColor: '#1C1C1F'
   }
 }