+++ packages/inventory/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @medusajs/inventory
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`98fe8fd00`](https://github.com/medusajs/medusa/commit/98fe8fd00a1912fde1d2a93b434bda500a213c14), [`2869763ea`](https://github.com/medusajs/medusa/commit/2869763ea9673cf5c5a3aa10b58f61b8e830ce21)]:
+  - @medusajs/medusa@1.7.15
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`902ed3c0b`](https://github.com/medusajs/medusa/commit/902ed3c0b21cd525d9975e59b7a8120ae5f7a895), [`fa4049cb5`](https://github.com/medusajs/medusa/commit/fa4049cb51232b2ed7091b1322c3fc14cd23e451)]:
+  - @medusajs/medusa@1.7.14
+
 ## 1.0.10
 
 ### Patch Changes